+++ src/trx/game/shell/config.c
@@ -191,7 +191,8 @@ void Shell_HandleConfigChange(const CONFIG_CHANGE *const change)
         Output_RefreshObjectMeshes();
     }
 
-    if (L_CHANGED(visuals.braid_status) || L_CHANGED(visuals.sunglasses_mode)) {
+    if (L_CHANGED(visuals.braid_status) || L_CHANGED(visuals.sunglasses_mode)
+        || L_CHANGED(visuals.sunglasses_reflective)) {
         Lara_Skin_ApplyOutfit();
     }
     if (L_CHANGED(visuals.lara_outfit) || L_CHANGED(visuals.golden_lara)) {